src/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
antiorm==1.2.1
attrs==22.1.0
cachelib==0.9.0
click==8.1.3
db==0.1.1
Flask==2.2.2
Flask-Caching==2.0.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restplus==0.13.0
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
psycopg2-binary==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.42
typing_extensions==4.4.0
webargs==8.2.0
Werkzeug==2.2.0
